added quick simple tests, updated datatype!, cleanup

+++ app/src/main/java/android/mobilechallenge/models/ExchangeRate.java
@@ -1,15 +1,17 @@
 package android.mobilechallenge.models;
 
+import java.math.BigDecimal;
+
 /**
  * ExchangeRate used for storing each exchange rate into DB and used in populating grid adapter.
  */
 public class ExchangeRate {
 
     private String currencyName;
-    private double exchangeRate;
-    private double convertedValue;
+    private BigDecimal exchangeRate;
+    private BigDecimal convertedValue;
 
-    public ExchangeRate(String currencyName, double exchangeRate) {
+    public ExchangeRate(String currencyName, BigDecimal exchangeRate) {
         this.currencyName = currencyName;
         this.exchangeRate = exchangeRate;
     }
@@ -24,21 +26,21 @@ public class ExchangeRate {
     /**
      * Used to get the exchange rate
      */
-    public double getExchangeRate() {
+    public BigDecimal getExchangeRate() {
         return this.exchangeRate;
     }
 
     /**
      * Used to set the converted value
      */
-    public void setConvertedValue(double value) {
+    public void setConvertedValue(BigDecimal value) {
         this.convertedValue = value;
     }
 
     /**
      * Used to get the converted value
      */
-    public double getConvertedValue() {
+    public BigDecimal getConvertedValue() {
         return this.convertedValue;
     }
 }
